chore: move failNotifyFunc to handler.go
